[feat] 채팅방 placeholder 컴포넌트 구현
